refactor(Step3) : 패키지 정리
