+++ src/main/java/racingGame/logic/logic/RacingGame.java
@@ -1,4 +1,4 @@
-package racingGame;
+package racingGame.logic.logic;
 
 public class RacingGame {
 	private static int RANDOM_RANGE = 10;
